2022.10.19(1)：P164 整合Sentinel作为熔断器

+++ service/service_edu/src/main/java/moe/tree/eduservice/service/impl/CourseServiceImpl.java
@@ -105,10 +105,22 @@ public class CourseServiceImpl extends ServiceImpl<CourseMapper, Course> impleme
 
 	@Override
 	public boolean removeCourse(String courseId) {
-		videoSerivce.removeVideoByCourseId(courseId);
-		chapterService.removeChapterByCourseId(courseId);
-		courseDescriptionService.removeById(courseId);
-		baseMapper.deleteById(courseId);
+		boolean res = videoSerivce.removeVideoByCourseId(courseId);
+		if(!res) {
+			return false;
+		}
+		res = chapterService.removeChapterByCourseId(courseId);
+		if(!res) {
+			return false;
+		}
+		res = courseDescriptionService.removeById(courseId);
+		if(!res) {
+			return false;
+		}
+		int cnt = baseMapper.deleteById(courseId);
+		if(cnt <= 0) {
+			return false;
+		}
 		return true;
 	}
 }
